+++ lib/Rule.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { shallow, mount } from 'enzyme';
+import { shallow } from 'enzyme';
 
-import Rule from '../lib/Rule';
-import { ActionElement, ValueSelector, ValueEditor } from '../lib/controls/index';
+import Rule from './Rule';
+import { ActionElement, ValueSelector, ValueEditor } from './controls/index';
 
 describe('<Rule />', ()=> {
     let controls, classNames, schema, props;
