+++ lib/RuleGroup.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { shallow, mount } from 'enzyme';
+import { shallow } from 'enzyme';
 
-import RuleGroup from '../lib/RuleGroup';
-import { ActionElement, ValueSelector } from '../lib/controls/index';
+import RuleGroup from './RuleGroup';
+import { ActionElement, ValueSelector } from './controls/index';
 
 describe('<RuleGroup />', ()=> {
     let controls, classNames, schema, props;
