+++ lib/controls/ActionElement.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { ActionElement } from '../../lib/controls/index';
+import { ActionElement } from './index';
 
 describe('<ActionElement />', ()=> {
