+++ lib/controls/ValueEditor.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { ValueEditor } from '../../lib/controls/index';
+import { ValueEditor } from './index';
 
 describe('<ValueEditor />', ()=> {
