+++ lib/controls/ValueSelector.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { ValueSelector } from '../../lib/controls/index';
+import { ValueSelector } from './index';
 
 describe('<ValueSelector />', ()=> {
